add setting feature

+++ database/migrations/2025_11_21_000007_create_order_items_table.php
@@ -11,16 +11,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('order_items', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('order_id')->constrained()->onDelete('cascade');
-            $table->foreignId('product_id')->constrained();
-            $table->foreignId('product_variant_id')->nullable()->constrained()->nullOnDelete();
-            $table->integer('quantity');
-            $table->decimal('unit_price', 15, 2);
-            $table->decimal('sub_total', 15, 2);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('order_items')) {
+            Schema::create('order_items', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('order_id')->constrained()->onDelete('cascade');
+                $table->foreignId('product_id')->constrained();
+                $table->foreignId('product_variant_id')->nullable()->constrained()->nullOnDelete();
+                $table->integer('quantity');
+                $table->decimal('unit_price', 15, 2);
+                $table->decimal('sub_total', 15, 2);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
